appwrite/discord-bots/public-bot/requirements.txt
discord.py
colorama
python-dotenv
meilisearch
discord